refactor: move tweakKey() to taproot utils
