+++ src/payments/taprootutils.d.ts
@@ -1,7 +1,6 @@
 /// <reference types="node" />
 import { Tapleaf, Taptree } from '../types';
 export declare const LEAF_VERSION_TAPSCRIPT = 192;
-export declare function rootHashFromPath(controlBlock: Buffer, leafHash: Buffer): Buffer;
 interface HashLeaf {
     hash: Buffer;
 }
@@ -10,6 +9,10 @@ interface HashBranch {
     left: HashTree;
     right: HashTree;
 }
+interface TweakedPublicKey {
+    parity: number;
+    x: Buffer;
+}
 /**
  * Binary tree representing leaf, branch, and root node hashes of a Taptree.
  * Each node contains a hash, and potentially left and right branch hashes.
@@ -17,6 +20,7 @@ interface HashBranch {
  * and calculating merkle inclusion proofs when constructing a control block.
  */
 export declare type HashTree = HashLeaf | HashBranch;
+export declare function rootHashFromPath(controlBlock: Buffer, leafHash: Buffer): Buffer;
 /**
  * Build a hash tree of merkle nodes from the scripts binary tree.
  * @param scriptTree - the tree of scripts to pairwise hash.
@@ -33,4 +37,5 @@ export declare function toHashTree(scriptTree: Taptree): HashTree;
 export declare function findScriptPath(node: HashTree, hash: Buffer): Buffer[] | undefined;
 export declare function tapleafHash(leaf: Tapleaf): Buffer;
 export declare function tapTweakHash(pubKey: Buffer, h: Buffer | undefined): Buffer;
+export declare function tweakKey(pubKey: Buffer, h: Buffer | undefined): TweakedPublicKey | null;
 export {};
